+++ lib/rails/assets/manifest/helper.rb
@@ -2,6 +2,26 @@
 
 module Rails::Assets::Manifest
   module Helper
+    URI_REGEXP = %r{^[-a-z]+://|^(?:cid|data):|^//}i.freeze
+
+    def path_to_asset(source, options)
+      if (entry = ::Rails::Assets::Manifest.lookup(path_with_extname(source, options)))
+        # Directly return the entry src if it is a fully qualified URL. Otherwise,
+        # Rails will join the URL with `relative_url_root` and/or asset host.
+        return entry.src if URI_REGEXP.match?(entry.src)
+
+        if entry.src[0] == '/'
+          # When the asset name starts with a slash, Rails will skip an
+          # additional lookup via `#compute_asset_path` and directly use the
+          # provided path. As we already have looked up the manifest entry here,
+          # we can pass the entry source, but only *if* it starts with a slash.
+          return super entry.src, options
+        end
+      end
+
+      super
+    end
+
     def compute_asset_path(name, _options)
       ::Rails::Assets::Manifest.lookup!(name).src
     rescue EntryMissing